+++ packages/bookmark/src/types/RenderBookmarkItemProps.ts
@@ -19,6 +19,7 @@ export interface RenderBookmarkItemProps {
     hasSubItems: boolean;
     index: number;
     isExpanded: boolean;
+    path: string;
     onClickItem: () => void;
     onClickTitle: () => void;
     onToggleSubItems: () => void;